backend/requirements.txt
Flask-SQLAlchemy
Flask-Migrate
flask
psycopg2-binary
SQLAlchemy
requests
pyserial